Add variadic display/calling of Janus functions.
Note that even pure functions have to have IO return types, because I get
segfaults in the presence of unsafePerformIO. This segfault is caused by dlclose
being called before the compiled function is called. I can't figure out why, and
this library is intended to do grungy imperative things under the hood anyhow,
so let's just slap IO on everything for the moment.